Allow the user to specify whether the TITokenFieldView should be scrollable.
If the user allows scrolling, we still disable scrolling internally while the results table
is showing - if the user allows scrolling (which is the default value) then there should
be *no change in behaviour* in this commit.

If the user disallows scrolling, then the entire view should *never* be scrollable. This
introduces the ability to specify add a scrollview to the contentView, meaning that the
To: box always appears on screen, and doesn't scroll away with the content, as is the
current behaviour.