When typing in the token field on iOS 7, the field would disappear behind the nav bar.
This is because the content insets were modified by the system to account for the transparent nav
bar (top inset = 64) but the contentOffset-changing code in setResultsModeEnabled:animated: assumed
that all changes would be relative to top inset = 0.

We can fix this simply by including the contentInsets.top in the calculation of the contentOffset's Y value.